Moved files into 'parse/' directory
