+++ Makefile
@@ -5,10 +5,11 @@ CC = @cc
 CFLAGS = -Wall -Wextra -Werror -g -fsanitize=address
 RM = rm -rf
 
-SRC_MSH = init.c main.c parse.c parse_split.c parse_utils_arr.c parse_utils.c
+SRC_MSH	=	init.c main.c 
+SRC_PRS	=	parse.c parse_split.c parse_utils_arr.c parse_utils.c
 
-SRCS = $(addprefix src/, $(SRC_MSH))
-OBJS = $(patsubst src/%, $(OBJS_DIR)/%, $(SRCS:%.c=%.o))
+SRCS = $(addprefix src/, $(SRC_MSH)) $(addprefix src/parse/, $(SRC_PRS))
+OBJS = $(addprefix, $(OBJS_DIR)/, $(SRC_MSH:%.c=%.o)) $(addprefix, $(OBJS_DIR)/, $(SRC_PRS:%.c=%.o))
 
 GREEN	=	"\033[0;32m"
 NC		=	"\033[0m"
