+++ src/parse/parse_utils.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
 int	str_words_quotes(t_var *var, char *str, int i)
 {
